+++ component_dev/alert/src/index.js
@@ -19,11 +19,22 @@ import yoConfirm from '../../confirm/src';
 
 /**
  * @method Alert
- * @param {String} content 组件显示的内容
+ * @description Alert API，调用以后在屏幕正中弹出一个Alert，可以按照option对象参数调用，也可以使用简易
+ * 调用方式如``Alert(content,title)``
+ * @param {Object} option 配置对象，里面可以接受如下属性：
+ * @param {String} [content] 组件显示的内容
  * @param {String} [title] 组件显示的标题
+ * @param {String} [btnText] 组件按钮的文本
  * @param {String | Object} [animation] 组件显隐执行的动画，格式同Dialog组件
- * @constructor Alert构造函数
+ * @constructor Alert API
  */
-export default function Alert(content = '', title = '', animation = 'fade') {
-    return yoConfirm(content, title, animation, false);
+export default function Alert(content = '', title = '', btnText = ['确定', ''], animation = 'fade') {
+    if (typeof content === 'object') {
+        const opt = content;
+        content = opt.content != null ? opt.content : '';
+        title = opt.title != null ? opt.title : '';
+        btnText = opt.btnText != null ? [opt.btnText, ''] : ['确定', ''];
+        animation = opt.animation || 'fade';
+    }
+    return yoConfirm(content, title, btnText, animation, false);
 }