+++ component_dev/swipemenulist/test/demo.js
@@ -5,11 +5,12 @@ import { getArrayByLength } from '../../common/util';
 import '../../common/touchEventSimulator';
 
 function getRandomList(size) {
-    return getArrayByLength(size).fill(1).map(num=>parseInt(Math.random() * 100));
+    return getArrayByLength(size).fill(1).map(num => parseInt(Math.random() * 100));
 }
 
+let guid = -1;
 function getRandomDataSource(size) {
-    return getRandomList(size).map(num=>({ text: num }));
+    return getRandomList(size).map(num => ({ key: ++guid, text: num }));
 }
 
 class SwipeMenuListDemo extends Component {
@@ -17,16 +18,16 @@ class SwipeMenuListDemo extends Component {
     constructor() {
         super();
         const self = this;
-        const testData = getRandomDataSource(100).map((item, i)=>({
+        const testData = getRandomDataSource(100).map((item, i) => ({
             ...item,
             id: i,
             randomHeight: parseInt(10 + Math.random() * 40, 10),
             action: [
                 {
-                    content: '删除',
-                    tap(item, i, swipeMenu){
+                    text: '删除',
+                    onTap(item, i, swipeMenu){
                         self.setState({
-                            dataSource: self.state.dataSource.filter((it, index)=>it.id !== item.id)
+                            dataSource: self.state.dataSource.filter((it, index) => it.id !== item.id)
                         });
                     }
                 }